Add AI Questions for All Rubrics (#6034)
@kiki-lee and I went through and tried to determine learning objectives for each tutorial, and then tried to encapsulate those into AI questions.

I tried to keep them to a single sentence / question on the hypothesis that this feature will be received better if teachers don't feel they need to construct a whole preamble to their question before asking it. If that turns out to be untrue, we can adjust, but we should also increase the text field to be multi-line, since it currently gets cut off.