Fix Build with Theming (#6108)
The new theming changes in pxt (microsoft/pxt#10391) removed some less variables and set others to css variables, which cannot be used in less color-manipulation functions (like lighten, darken, etc...). As a result, the build for microbit broke when referencing newer versions of pxt (or local).

In this change, I fix those css build errors, add a default theme, and update themepack variables to reference new theme variables. In time, we may want to fully swap over from themepack to this, but this is a stop-gap until we're focused on micro:bit again.

This isn't meant to look perfect. I mostly want to make sure everything can still build. Making it look nice is secondary, and more likely something we'll focus on when it comes time for our next micro:bit release.