Microbit - Teacher Tool Theme Fixes (#6132)
In microsoft/pxt#10481, I have swapped out the themepack with our new theme variables, so we can remove the themepack file from microbit. The one thing we needed to keep was the font variable, so I've moved that into style.less instead.

I've also added a color that was missed in the original theme file. This already existed in arcade themes (and the base theme).