Add Dark Theme (and fix a few theming bugs) (#6343)
This adds a dark theme for micro:bit and fixes a few theming bugs along the way. While doing this work, I found a few places in the micro:bit styles.less that were hard-coded to light-theme-specific values. I themed them when I could and moved the rest into overrides.