Remove white background from high contrast sim LEDs (#6418)
Previously, we used a white background for the sim LEDs in high contrast mode, I suspect so that it contrasted with the unlit LEDs as well as the lit ones. But it looks a little like a bug...instead of doing it that way, we can use the normal red-on-black appearance (with slightly brighter red) and outline the unlit LEDs with white.